+++ frontend/src/pages/Traces/TraceLogs.tsx
@@ -51,14 +51,17 @@ export const TraceLogs: React.FC = () => {
 	})
 
 	const fetchMoreWhenScrolled = React.useCallback(
-		(containerRefElement?: HTMLDivElement | null) => {
+		(
+			containerRefElement?: HTMLDivElement | null,
+			disableBackwards?: boolean,
+		) => {
 			if (containerRefElement) {
 				const { scrollHeight, scrollTop, clientHeight } =
 					containerRefElement
 
 				if (scrollHeight - scrollTop - clientHeight < 100) {
 					fetchMoreForward()
-				} else if (scrollTop === 0) {
+				} else if (!disableBackwards && scrollTop === 0) {
 					fetchMoreBackward()
 				}
 			}
@@ -91,7 +94,6 @@ export const TraceLogs: React.FC = () => {
 				justifyContent="stretch"
 				display="flex"
 				overflow="hidden"
-				maxHeight="full"
 			>
 				<Box
 					borderRadius="6"
@@ -130,6 +132,7 @@ export const TraceLogs: React.FC = () => {
 								query={query}
 								selectedCursor={undefined}
 								fetchMoreWhenScrolled={fetchMoreWhenScrolled}
+								bodyHeight="400px"
 							/>
 						)}
 					</Box>