+++ e2e/express-ts/CHANGELOG.md
@@ -1,5 +1,13 @@
 # e2e-express-ts
 
+## 1.0.2
+
+### Patch Changes
+
+-   Updated dependencies [bd5021a6c]
+-   Updated dependencies [2fce77c13]
+    -   @highlight-run/node@3.6.5
+
 ## 1.0.1
 
 ### Patch Changes